cobalt: Consolidate WebContents logic into Shell (#8296)
Remove CobaltWebContentsDelegate and CobaltWebContentsObserver classes.
Their responsibilities for WebContents interaction and lifecycle
management are now handled directly by the content::Shell class.

This refactoring simplifies the Cobalt browser architecture by reducing
intermediary classes. It centralizes all WebContents-related logic
within the Shell, which already inherits WebContentsDelegate and
WebContentsObserver, making it a more cohesive design. This includes
JavaScript injection, storage migration, media access permission,
OOM intervention setup, and application closing.

Issue: 466120031
(cherry picked from commit 13c80ea1b5c075c820be462c4bd77692c91ebcc6)
