+++ cobalt/browser/cobalt_content_browser_client.h
@@ -17,8 +17,6 @@
 
 #include "base/threading/thread_checker.h"
 #include "cobalt/browser/client_hint_headers/cobalt_trusted_url_loader_header_client.h"
-#include "cobalt/browser/cobalt_web_contents_delegate.h"
-#include "cobalt/browser/cobalt_web_contents_observer.h"
 #include "cobalt/shell/browser/shell_content_browser_client.h"
 #include "content/public/browser/generated_code_cache_settings.h"
 #include "mojo/public/cpp/bindings/pending_receiver.h"
@@ -133,7 +131,6 @@ class CobaltContentBrowserClient : public content::ShellContentBrowserClient {
   void CreateVideoGeometrySetterService();
 
   std::unique_ptr<CobaltWebContentsObserver> web_contents_observer_;
-  std::unique_ptr<CobaltWebContentsDelegate> web_contents_delegate_;
   std::unique_ptr<media::VideoGeometrySetterService, base::OnTaskRunnerDeleter>
       video_geometry_setter_service_;
