+++ cobalt/browser/cobalt_web_contents_observer.h
@@ -15,26 +15,26 @@
 #ifndef COBALT_BROWSER_COBALT_WEB_CONTENTS_OBSERVER_H_
 #define COBALT_BROWSER_COBALT_WEB_CONTENTS_OBSERVER_H_
 
-#include "cobalt/shell/browser/shell_content_browser_client.h"
 #include "content/public/browser/web_contents_observer.h"
 
 #include "components/js_injection/browser/js_communication_host.h"
 
 namespace cobalt {
 
+// This class is used to observe WebContents lifecycles within
+// CobaltContentBrowserClient. Its primary purpose is to dispatch web events.
+// Most WebContentsObserver functionalities should be implemented in
+// cobalt/shell/browser/shell.{h,cc} instead. This class should only contain
+// logic essential for the browser client's core responsibilities.
 class CobaltWebContentsObserver : public content::WebContentsObserver {
  public:
   CobaltWebContentsObserver(content::WebContents* web_contents);
 
-  void PrimaryMainDocumentElementAvailable() override;
+  CobaltWebContentsObserver(const CobaltWebContentsObserver&) = delete;
+  CobaltWebContentsObserver& operator=(const CobaltWebContentsObserver&) =
+      delete;
 
-  void DidFinishNavigation(
-      content::NavigationHandle* navigation_handle) override;
-
- private:
-  void RegisterInjectedJavaScript();
-
-  std::unique_ptr<js_injection::JsCommunicationHost> js_communication_host_;
+  ~CobaltWebContentsObserver() override;
 };
 
 }  // namespace cobalt
