+++ cobalt/shell/embedded_resources/cobalt_java_script_polyfill/h5vcc_platform_service_on_java_bridge.js
