+++ cobalt/shell/embedded_resources/loader_embedded_resources/splash.html
